renames and import / export fixes
