+++ libs/newsletters-data-client/src/index.ts
@@ -4,12 +4,14 @@ export * from './lib/deriveNewsletterFields';
 export * from './lib/draft-service';
 export * from './lib/draft-storage';
 export * from './lib/draft-to-newsletter';
-export * from './lib/schemas/emailEmbedSchema';
+export * from './lib/schemas/email-embed-data-type';
 export * from './lib/launch-service';
 export * from './lib/schemas/legacy-newsletter-type';
 export * from './lib/schemas/meta-data-type';
 export * from './lib/schemas/newsletter-data-type';
 export * from './lib/schemas/data-collection-schema';
+export * from './lib/schemas/rendering-options-data-type';
+export * from './lib/schemas/theme-enum-data-type';
 export * from './lib/newsletter-storage';
 export * from './lib/storage-response-types';
 export * from './lib/transformDataToLegacyNewsletter';
