+++ libs/newsletters-data-client/src/lib/schemas/email-embed-data-type.ts
